Add timer to public index
Thanks to *anonymous* for one of the best feedback ever:

Be careful with visual "alignment problems". (see link) You don't have
any on the homepage but on the timer page, because there's so much white
space and no visual containers the eye has trouble lining things up.
http://www.chrisnorstrom.com/2011/05/quick-fix-up-redesign-of-tap11-com/
The big ass logo on the home page is unnecessary, instead use that space
to explain to me WHAT the pomodoro technique is and why i should use
your site and how easy it is to sign up. The timer should be inside of
some kind of light grey container so it's better visually organized.
There needs to be some more space beneath the timer to separate it from
the "leaderboards". Change "squash" to "reset timer" and don't use a
javascript confirm box. Creative wording of buttons is bad practice and
confuses users. I really don't need an account for this, there should
be a guest pass option. Or hell, load the timer up on the home page and
have "log in / sign up to track my progress" optional.