Add API tomatoes controller
The new controller implements only `#show` and `#index` actions at
the moment.

Get the current user's list of tomatoes:

    GET /api/tomatoes?token=123
    GET /api/tomatoes?token=123&page=2

Get a current user's tomato:

    GET /api/tomatoes/123?token=123

The `#index` action returns a paginated list of tomatoes, but it
doesn't include meta information such the page number and the total
number of pages, yet.

The initial JSON representation of a tomato resource is:

    {
      "id": "57f9c9187c8402dd306d1c88",
      "created_at": "2016-10-09T04:35:36.952Z",
      "updated_at": "2016-10-09T04:35:36.952Z",
      "tags": ["one", "two"]
    }