Add Api::ProjectsController
The `Api::ProjectsController` endpoints behavior is very similar to
`Api::TomatoesController` endpoints behavior.

Actions:

* `GET #index`
* `GET #show`
* `POST #create`
* `PATCH/PUT #update`
* `DELETE #destroy`

Get the current user's list of projects:

    GET /api/projects?token=123
    GET /api/projects?token=123&page=2

Get a current user's project:

    GET /api/projects/123?token=123

Create a new project:

    POST /api/projects

```json
{
  "token": "123",
  "project": {
    "name": "Web app",
    "tag_list": "one, two",
    "money_budget": 1200,
    "time_budget": 120
  }
}
```

* `name` is a requested attribute.
* `money_budget` is an integer value that represents the amount
  of money allocated for the project expressed in user's default
  currency
* `time_budget` is an integer value that represents the amount of
  time estimated for the project expressed in hours

Update an existing project:

    PATCH /api/tomatoes/123

```json
{
  "token": "123",
  "project": {
    "name": "Web app updated",
    "tag_list": "one, two"
  }
}
```

Delete an existing project:

    DELETE /api/projects/123?token=123

The `#index` action returns a paginated list of projects, it includes
pagination information such the current page, the total number of
pages, and the total number of resources:

```json
{
  "projects": [
    {
      "id": "57f9c9377c8402dd306d1c8b",
      "name": "Web app",
      "created_at": "2016-10-09T04:36:07.787Z",
      "updated_at": "2016-10-09T04:36:07.787Z",
      "tags": [
        "one",
        "two"
      ],
      "money_budget": 1200,
      "time_budget": 120
    }
  ],
  "pagination": {
    "current_page": 1,
    "total_pages": 1,
    "total_count": 1
  }
}
```

The JSON representation of a project resource is:

{
  "id": "57f9c9377c8402dd306d1c8b",
  "name": "Web app",
  "created_at": "2016-10-09T04:36:07.787Z",
  "updated_at": "2016-10-09T04:36:07.787Z",
  "tags": [
    "one",
    "two"
  ],
  "money_budget": 1200,
  "time_budget": 120
}
```

A request to `POST #create` creates a new project associated to the
current user. It returns `201 Created` and a JSON representation of
the resource created in case of success. It returns `422 Unprocessable
Entity` and a JSON representation of the validation errors in case of
failure.

Example response body (failure):

```json
{
  "name": [
    "can't be blank"
  ]
}
```

A request to `PATCH #update` updates an existing project associated to
the current user. It returns `200 Ok` and a JSON representation of the
resource created in case of success. It returns `422 Unprocessable
Entity` and a JSON representation of the validation errors in case of
failure. See `POST #create` to see a failed operation response example.

A request to `DELETE #destroy` removes an existing tomato associated to
the current user. It returns `204 No Content` in case of success.