Rescue in case of duplicate key error
If two threads try to concurrently create a score for the same period
and the same user, one of them will fail because of the `unique`
constraint on the `uid` index. The one that fails should retry the
`add_delta` operation to apply the delta to the record that is now
present in the collection.

I don't know exactly how can I test this race condition. Do you have
any suggestion @bugant/@dalpo?

I tried manually to create two `DailyScore` records with the same
`uid` and I received the error:

    Mongo::Error::OperationFailure:
      E11000 duplicate key error collection:
      tomatoes_app_development.daily_scores index: uid_index dup
      key: { : 1 } (11000)

That's the reason why I'm rescuing from `Error::OperationFailure`. I
don't know if there's a more appropriate error I could rescue from.
It seems like `Error::OperationFailure` is too generic. If that's the
case we could be retrying indefinitely on an error that is not a
duplicate key error. Again I'm open to suggestions.