Add API leaderboards endpoints
This commit adds 4 new endpoints to the API:

1. GET /api/leaderboard/daily
2. GET /api/leaderboard/weekly
3. GET /api/leaderboard/monthly
4. GET /api/leaderboard/overall

All these endpoints return a paginated list and accept a `page` param
to navigate through the available pages.

Leaderboard responses will be all similar and will include a `user`,
with basic user information (`id`, `name`, `image`), and a `score`,
that is the number of tomatoes completed by the user in the selected
period.

Example response:

    {
      "scores": [
        {
          "user": {
            "id": "586c2be37c8402b5d1db406c",
            "name": "Giovanni Cappellotto",
            "image": "http://tomato.es/images/users/586c2be37c8402b5d1db406c.png"
          },
          "score": 2
        }
      ],
      "pagination": {
        "current_page": 1,
        "total_pages": 1,
        "total_count": 1
      }
    }