Remind if pomodoro has not been saved or started
`TTReminder` hooks on three timer events: `timer_start`, `timer_stop`,
`new_tomato_form`. It reminds the user if the current tomato has not
been saved or if a new one could be started after a break.

The script will start notifying the user after 10 + 6^1 seconds. It will
send a second notification after 10 + 6^2 seconds from the first one. It
stops after the fourth notification, that should arrive after 10 + 6^4
seconds from the previous notification.

Reminder notifications times:

1. 10 + 6^1 = 16 seconds after the event
2. 10 + 6^2 = 46 seconds after the 1st notification
3. 10 + 6^3 = 3 minutes, 46 seconds after the 2nd notification
4. 10 + 6^4 = 21 minutes, 46 seconds after the 3rd notification

This is the first approach that I though to fix
https://github.com/tomatoes-app/tomatoes/issues/109.

An alternative approach would be to notify the user only once, but to
make the notification last longer. This would prevent the case when the
user just doesn't want to start a new tomato, but he/she will keep
receiving notifications.

I didn't implement a sound notification, as suggested by
https://github.com/tomatoes-app/tomatoes/issues/93, because I think it
would be too *noisy*.