Update Notifier.notify interface
I think `Notification` interface changed or this method implementation
has always been wrong because the first parameter is the notification
title, and the body should be included in the options object, not the
other way around.

Adding also `requireInteraction` that will be useful later and removing
some spare `console.log`s.