Fix invalid time zone argument error
If a user has an empty string set as a time zone, API requests will fail
with the error:

    An ArgumentError occurred in tomatoes#create:
      Invalid Timezone

This update fixes this issue and a more generic issue if the user has a
time zone that is an invalid time zone. Note: time zone attribute
validation is missing, maybe we should validate that values for that
attribute are included in the list of valid time zone names.