Fix NoMethodError
If a POST request includes a `tomato` parameter that is not a hash,
Rails raises an exception because the value doesn't define `#permit`.

Fix the issue by *trying* to send the `#permit` message to the param
value.

Note: this isn't an optimal solution because malformed requests will
end up creating tomatoes, but it's better than returning a 500
response.