Fix routing error for tags#show
Tag ids can include any character, but by default Rails' router
can't parse `.` characters as part of an id. This caused an error
that prevented some tags to be showed. This change includes a
regression test and a fix for this issue, that is a new routing
rule that uses a different pattern to match. See also:
http://guides.rubyonrails.org/routing.html#route-globbing-and-wildcard-segments