Dockerize project
This commit adds Docker support with updated documentation on how to
contribute to the project.

This change is retro-compatible. You are not forced to run Docker and
you can keep working with your current setup. But if you prefer to use
containers or if you are a new contributor should help setting up the
project.